requirements.txt
pyautogui==0.9.54
pynput==1.7.6
Pillow==10.1